Split into class files and separated the MQTT service
